added joystick support
